♻️ Refactor CLI architecture for better maintainability
- Split monolithic 393-line cli.ts into modular structure
- Extract commands into separate files (init, compile, preview, help)
- Create reusable lib modules (file-utils, rule-generator)
- Consolidate types and constants
- Reduce main CLI file from 393 to 28 lines
- Improve code organization and testability
